fix: use correct type for none input

+++ src/TesiraDspRouter.cs
@@ -140,7 +140,7 @@ namespace Pepperdash.Essentials.Plugins.DSP.Biamp.Tesira
                 return;
             }
 
-            InputPorts.Add(new RoutingInputPort("None", eRoutingSignalType.Audio, eRoutingPortConnectionType.BackplaneOnly, 0, this));
+            InputPorts.Add(new RoutingInputPort("None", eRoutingSignalType.Audio, eRoutingPortConnectionType.BackplaneOnly, (uint)0, this));
 
             foreach (
                 var input in
